second

+++ 1611.08408/utils.py
@@ -5,7 +5,7 @@ import cv2
 
 def interp(src, zoom=None, shrink=None):
     shape = src.size()
-    src_np = src.numpy().astype(np.uint8)
+    src_np = src.numpy()
     dst_np = np.zeros(shape=shape)
 
     if zoom is not None:
@@ -16,7 +16,7 @@ def interp(src, zoom=None, shrink=None):
         width_out = (shape[1] - 1) / shrink + 1
 
     for index in xrange(shape[0]):
-        single = src_np[index, :, :, :]
+        single = src_np[index, :, :, :].astype(np.uint8)
         dst_np[index, :, :, :] = cv2.resize(single, (height_out, width_out), cv2.INTER_LINEAR)
     return torch.from_numpy(dst_np)
